Sorting not completed

+++ cviceni_20180220/cviceni_20180220/Classes/Debt.cs
@@ -13,6 +13,6 @@ namespace cviceni_20180220
         public int ID { get; set; }
         public int IDItem { get; set; }
         public DateTime DateToPay { get; set; }
-        public int RaisePercentage { get; set; }
+        public double RaisePercentage { get; set; }
     }
 }
